Refactor tooltip and favorite button to separate components
